+++ CHANGES.md
@@ -57,6 +57,12 @@
 * Added `relative_magnitudes` argument to 
   `eqcorrscan.core.match_filter.party.Party.lag_calc` to provide an in-flow
   way to compute relative magnitudes for detected events.
+* Events constructed from detections now include estimated origins alongside
+  the picks. These origins are time-shifted versions of the template origin and
+  should be used with caution. They are corrected for prepick.
+* Picks in detection.event are now corrected for prepick *if* the template is
+  given. This is now standard in all Tribe, Party and Family methods. Picks will
+  not be corrected for prepick in match_filter.
 * Fix #298 where the header was repeated in detection csv files. Also added
   a `write_detections` function to `eqcorrscan.core.match_filter.detection`
   to streamline writing detections.